+++ app/Policies/CertificatePolicy.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Policies;
+
+class CertificatePolicy extends AdminOnlyPolicy {}